requirements.txt
pyyaml
prettytable
tqdm